requirements.txt
# pyuavcan v0.1
pyserial
uavcan

# common
coloredlogs